# openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Dato API
  description: Dato API Demo
  version: '2.0'
servers:
  - url: http://localhost:8082
    description: Generated server url
...
paths:
...
  /api/v1/dato/{id}:
    get:
      tags:
        - dato
      summary: show a dato object by id
      description: show a dato object by identifier
      operationId: showDatoById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Dato'
            application/json:
              schema:
                $ref: '#/components/schemas/Dato'
        '404':
          description: Invalid input
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    put:
      tags:
        - dato
      summary: update a dato object by id
      description: update a dato object by identifier
      operationId: editDatoById
      parameters:
        - name: id
          in: path
          description: id user object
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatoDTO'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Dato'
            application/json:
              schema:
                $ref: '#/components/schemas/Dato'
        '404':
          description: Invalid input
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      tags:
        - dato
      summary: delete a dato object by id
      description: delete a dato object by identifier
      operationId: deleteDatoById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Dato'
            application/json:
              schema:
                $ref: '#/components/schemas/Dato'
        '404':
          description: Object not found
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /api/v1/dato/response:
    get:
      tags:
        - dato
      operationId: indexResponse_1
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dato'
    post:
      tags:
        - dato
      operationId: addDatoResponse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatoDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/Dato'
  /api/v1/dato/:
    get:
      tags:
        - dato
      summary: show list of dato objects
      description: Shows a list of dato in an output array
      operationId: index_6
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dato'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dato'
    post:
      tags:
        - dato
      summary: add a dato object
      description: add a dato object with a string input of 80 max characters
      operationId: addDato
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatoDTO'
        required: true
      responses:
        '200':
          description: Dato added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dato'
            application/xml:
              schema:
                $ref: '#/components/schemas/Dato'
...
components:
  schemas:
...
    DatoDTO:
      required:
        - cadena
      type: object
      properties:
        cadena:
          type: string
      description: Created user object
    ErrorMessage:
      required:
        - statusCode
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        description:
          type: string
    Dato:
      required:
        - cadena
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        cadena:
          type: string
...